Fix time drilling props
